# openapi.yaml
openapi: 3.0.0
info:
  title: WhatsApp Automation Bot API
  description: Production-ready WhatsApp automation bot system with API-like features powered by Playwright
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://your-domain.com/api
    description: Production server
security:
  - ApiKeyAuth: []
  - BearerAuth: []
...
paths:
...
  /queue:
    get:
      tags:
        - Queue
      summary: Get queue jobs
      description: Get queue jobs with pagination and filtering
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/StatusParam'
        - $ref: '#/components/parameters/JobTypeParam'
      responses:
        '200':
          description: Queue jobs retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/QueueJob'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /queue/stats:
    get:
      tags:
        - Queue
      summary: Get queue statistics
      description: Get queue processing statistics
      responses:
        '200':
          description: Queue statistics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      counts:
                        type: object
                        properties:
                          waiting:
                            type: integer
                          active:
                            type: integer
                          completed:
                            type: integer
                          failed:
                            type: integer
                      waiting:
                        type: integer
                      active:
                        type: integer
                      completed:
                        type: integer
                      failed:
                        type: integer
                      recentFailed:
                        type: array
                        items:
                          $ref: '#/components/schemas/FailedJob'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /queue/{id}/retry:
    post:
      tags:
        - Queue
      summary: Retry failed job
      description: Retry a failed queue job
      parameters:
        - name: id
          in: path
          required: true
          description: Job ID
          schema:
            type: string
      responses:
        '200':
          description: Job retried successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Job retried successfully"
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /queue/{id}/cancel:
    post:
      tags:
        - Queue
      summary: Cancel job
      description: Cancel a queue job
      parameters:
        - name: id
          in: path
          required: true
          description: Job ID
          schema:
            type: string
      responses:
        '200':
          description: Job cancelled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Job cancelled successfully"
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /queue/pause:
    post:
      tags:
        - Queue
      summary: Pause queue
      description: Pause queue processing
      responses:
        '200':
          description: Queue paused successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Queue paused successfully"
        '401':
          $ref: '#/components/responses/Unauthorized'
  /queue/resume:
    post:
      tags:
        - Queue
      summary: Resume queue
      description: Resume queue processing
      responses:
        '200':
          description: Queue resumed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Queue resumed successfully"
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /templates:
    post:
      tags:
        - Templates
      summary: Create template
      description: Create a new message template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - content
              properties:
                name:
                  type: string
                  description: Template name
                  example: "Welcome Message"
                content:
                  type: string
                  description: Template content
                  example: "Hello {name}, welcome to our service!"
                description:
                  type: string
                  description: Template description
                  example: "Welcome message for new users"
      responses:
        '201':
          description: Template created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Template created successfully"
                  templateId:
                    type: integer
                    example: 1
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /templates/{id}:
    delete:
      tags:
        - Templates
      summary: Delete template
      description: Delete a template
      parameters:
        - name: id
          in: path
          required: true
          description: Template ID
          schema:
            type: integer
      responses:
        '200':
          description: Template deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Template deleted successfully"
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Number of items per page
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
...
    StatusParam:
      name: status
      in: query
      description: Filter by job status
      required: false
      schema:
        type: string
        enum: [pending, active, completed, failed, cancelled]
    JobTypeParam:
      name: jobType
      in: query
      description: Filter by job type
      required: false
      schema:
        type: string
        enum: [send_message, send_media, send_template]
  schemas:
...
    QueueJob:
      type: object
      properties:
        id:
          type: integer
          example: 1
        job_id:
          type: string
          example: "550e8400-e29b-41d4-a716-446655440000"
        job_type:
          type: string
          enum: [send_message, send_media, send_template]
          example: "send_message"
        status:
          type: string
          enum: [pending, active, completed, failed, cancelled]
          example: "pending"
        recipient:
          type: string
          example: "1234567890"
        content:
          type: string
          example: "Hello from WhatsApp Bot!"
        media_url:
          type: string
          format: uri
        media_type:
          type: string
        retry_count:
          type: integer
          example: 0
        error_message:
          type: string
        created_at:
          type: string
          format: date-time
          example: "2024-01-01T12:00:00.000Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-01-01T12:00:00.000Z"
    FailedJob:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        failedReason:
          type: string
        timestamp:
          type: integer
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 50
        total:
          type: integer
          example: 1000
        totalPages:
          type: integer
          example: 20
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "Validation failed"
              details:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                    message:
                      type: string
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "Invalid or expired token"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "Resource not found"
...